requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.1
line-bot-sdk==3.9.0
ngrok==1.0.0
flask-openapi3==3.0.2
google-cloud-aiplatform==1.43.0
opencv-python==4.9.0.80
schedule==1.2.1
